[IMP] Moving simplify_path to utils
