+++ branches.py
@@ -11,6 +11,8 @@ import json
 import logging
 import argparse
 
+from lib.utils import simplify_path
+
 logging.basicConfig(level=logging.DEBUG,
     format='%(asctime)s - %(name)s - %(levelname)s - %(message)s')
 logger = logging.getLogger('info_branchs')
@@ -33,30 +35,6 @@ if not path:
     path = os.getcwd()
 save = args.save
 load = args.load
-
-
-def simplify_path(b_info):
-    """This function deletes all common directories in branches' path
-
-    :param b_info: List of dictionaries with branches' info
-    :return: List of dictionaries with branches' info
-    """
-    logger.debug("Deleting all common branches' path")
-    repeated = True
-    while repeated:
-        piece_path = []
-        for branch in b_info:
-            piece_path.append(branch['path'].split('/', 1)[0])
-        word = piece_path[0]
-        repeated = True
-        for each in piece_path:
-            if each != word:
-                repeated = False
-        if repeated:
-            for branch in b_info:
-                branch.update({'path': branch['path'].split('/', 1)[1]})
-    logger.debug("Common paths deleted")
-    return b_info
 
 
 def get_all_branches_info(path):
